--- src/main/java/lotto/utils/LottoValidator.java
@@ -1,29 +1,0 @@
-package lotto.utils;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-public class LottoValidator implements Validator<List<Integer>> {
-    private static final int MIN_NUMBER = 1;
-    private static final int MAX_NUMBER = 45;
-    private static final int LOTTO_NUMBER_COUNT = 6;
-
-    @Override
-    public void validate(List<Integer> numbers) {
-        if (numbers.size() != LOTTO_NUMBER_COUNT) {
-            throw new IllegalArgumentException("로또 번호는 6개의 숫자여야 합니다.");
-        }
-
-        Set<Integer> uniqueNumbers = new HashSet<>(numbers);
-        if (uniqueNumbers.size() != LOTTO_NUMBER_COUNT) {
-            throw new IllegalArgumentException("로또 번호에는 중복된 숫자가 없어야 합니다.");
-        }
-
-        for (int number : numbers) {
-            if (number < MIN_NUMBER || number > MAX_NUMBER) {
-                throw new IllegalArgumentException("로또 번호는 1부터 45 사이의 숫자여야 합니다.");
-            }
-        }
-    }
-}
