+++ src/main/java/lotto/utils/validator/Validator.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.utils.validator;
 
 public interface Validator<T> {
     void validate(T value);
